+++ media/media-picker/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/media-picker
 
+## 54.2.2
+
+### Patch Changes
+
+- [`81c6a2fcb2`](https://bitbucket.org/atlassian/atlassian-frontend/commits/81c6a2fcb2) - Fixed potential exception when synchronously accessing uninitialized RxJS subscription.
+
 ## 54.2.1
 
 ### Patch Changes